14 anki 2.1.51

+++ smr/exportsync.py
@@ -1,42 +1,45 @@
 import json
+from zipfile import ZipFile, ZIP_DEFLATED
 
-import aqt.main as sync
+import aqt
 from aqt.utils import tooltip
 
-from anki.utils import splitFields
+from anki.utils import split_fields
 
+from .dto.deckselectiondialoguserinputsdto import DeckSelectionDialogUserInputsDTO
 from .utils import *
 from .xminder import XmindImporter
 
 
 class MapSyncer:
 
-    def __init__(self, mw):
-        self.mw = mw
+    def __init__(self):
         self.notes2Sync = []
         self.srcDir = None
         self.tagList = None
-        self.mediaDir = re.sub(r"(?i)\.(anki2)$", ".media", self.mw.col.path)
+        self.mediaDir = re.sub(r"(?i)\.(anki2)$", ".media", aqt.mw.col.path)
         self.manifest = None
         self.fileBin = None
 
     def syncMaps(self):
-        self.mw.progress.start(immediate=True,
-                               label='processing SMR changes...')
-        self.mw.app.processEvents()
+        aqt.mw.progress.start(immediate=True,
+                              label='processing SMR changes...')
+        aqt.mw.app.processEvents()
         self.getNotes2Sync()
         docs2Sync = set(map(lambda n: n['meta']['path'], self.notes2Sync))
         for doc2Sync in docs2Sync:
             self.syncDoc(doc2Sync)
-        self.mw.progress.finish()
+
+        aqt.mw.col.tags.clear_unused_tags()
+        aqt.mw.progress.finish()
 
     def getNotes2Sync(self):
         """gets all notes with fields that were changed after their last import"""
-        xMid = xModelId(self.mw.col)
-        xNotes = list(self.mw.col.db.execute(
+        xMid = xModelId(aqt.mw.col)
+        xNotes = list(aqt.mw.col.db.execute(
             "select id, mod, flds from notes where mid = %s" % xMid))
         for xNote in xNotes:
-            fields = splitFields(xNote[2])
+            fields = split_fields(xNote[2])
             meta = json.loads(fields[23])
             # If the last time the note was edited was at least 10 Seconds after it was imported
             if xNote[1] > (meta['lastSync'] + 10):
@@ -44,10 +47,10 @@ class MapSyncer:
                     dict(meta=meta, fields=fields, nid=xNote[0]))
 
     def syncDoc(self, docPath):
-        self.mw.progress.update(
+        aqt.mw.progress.update(
             label="synchronizing %s" % os.path.basename(docPath),
             maybeShow=False)
-        self.mw.app.processEvents()
+        aqt.mw.app.processEvents()
         # create temp dir
         self.srcDir = tempfile.mkdtemp()
         self.fileBin = []
@@ -58,7 +61,7 @@ class MapSyncer:
         except FileNotFoundError:
             log = 'File "%s" not found, changes in "%s" not exported.' % (
                 docPath, os.path.basename(docPath))
-            tooltip(msg=log, period=6000, parent=self.mw)
+            tooltip(msg=log, period=6000, parent=aqt.mw)
             return
         content = xZip.read('content.xml')
         manifestContent = xZip.read("META-INF/manifest.xml")
@@ -69,28 +72,33 @@ class MapSyncer:
         sheets2Sync = set(map(lambda n: n['meta']['sheetId'], notes4Doc))
         sheets2Sync = list(
             map(lambda s: soup.find('sheet', {'id': s}), sheets2Sync))
-        for note in notes4Doc:
-            self.syncNote(note)
-        self.updateZip(docPath, 'content.xml', str(soup))
-        # Remove temp dir and its files
-        shutil.rmtree(self.srcDir)
-        # import sheets again
-        importer = XmindImporter(col=self.mw.col, file=docPath)
-        sheetImports = []
-        for sheet in sheets2Sync:
+        if len(sheets2Sync) > 0:
+            sheet = sheets2Sync[0]
+            for note in notes4Doc:
+                self.syncNote(note)
+            self.update_zip(docPath, 'content.xml', str(soup))
+            # Remove temp dir and its files
+            shutil.rmtree(self.srcDir)
+            # import sheets again
+            importer = XmindImporter(col=aqt.mw.col, file=docPath)
+            sheetImports = []
             print('importing sheet')
-            tag4Sheet = sum(set(self.mw.col.db.execute(
+            tag4Sheet = next(taglist[0].strip() for taglist in aqt.mw.col.db.execute(
                 "select tags from notes where flds like '%\"sheetId\": \"" +
-                sheet['id'] + "\"%'")), ())[0]
-            sheetNid = sum(set(self.mw.col.db.execute(
+                sheet['id'] + "\"%'"))
+            sheetNid = next(nid_list[0] for nid_list in aqt.mw.col.db.execute(
                 "select id from notes where flds like '%\"sheetId\": \"" +
-                sheet['id'] + "\"%'")), ())[0]
-            did4Sheet = sum(set(self.mw.col.db.execute(
-                "select did from cards where nid = %s" % sheetNid)), ())[0]
-            sheetImport = dict(sheet=sheet, tag=tag4Sheet, deckId=did4Sheet,
-                               repair=False)
-            sheetImports.append(sheetImport)
-        importer.importSheets(sheetImports)
+                sheet['id'] + "\"%'"))
+            did4Sheet = next(did_list[0] for did_list in aqt.mw.col.db.execute(
+                "select did from cards where nid = %s" % sheetNid))
+            user_inputs = DeckSelectionDialogUserInputsDTO(
+                deck_id=did4Sheet,
+                repair=False,
+                deck_name=re.search(r'^([^:]*)::', tag4Sheet).group(1)
+            )
+            importer.importSheets(user_inputs=user_inputs)
+            log = "\n".join(importer.log)
+            tooltip(log)
 
     def syncNote(self, note):
         print('synchronizing note')
@@ -161,48 +169,32 @@ class MapSyncer:
         imgTag = tag.find('xhtml:img')
         imgTag['xhtml:src'] = 'xap:' + newFullPath
 
-        # TODO: add code for changing images
-
-    def updateZip(self, zipname, filename, data):
-        """ taken from https://stackoverflow.com/questions/25738523/how-to-update-one-file-inside-zip-file-using-python, replaces one file in a zipfile"""
+    def update_zip(self, zipname, filename, data) -> None:
+        """
+        - replaces the content.xml file in the xmind file with the manager's content soup
+        - replaces the manifest.xml with the manager's manifest soup
+        - removes all files in the file_bin from the xmind file
+        - adds all files in files_2_add to the xmind file
+        code was adopted from
+        https://stackoverflow.com/questions/25738523/how-to-update-one-file-inside-zip-file-using-python,
+        """
         # generate a temp file
-        tmpfd, tmpname = tempfile.mkstemp(dir=os.path.dirname(zipname))
-        os.close(tmpfd)
-
-        # create a temp copy of the archive without filename
-        with zipfile.ZipFile(zipname, 'r') as zin:
-            with zipfile.ZipFile(tmpname, 'w') as zout:
-                zout.comment = zin.comment  # preserve the comment
-                for item in zin.infolist():
-                    if item.filename not in [filename,
-                                             'META-INF/manifest.xml'] + \
-                            self.fileBin:
-                        zout.writestr(item, zin.read(item.filename))
-
-        # replace with the temp archive
+        temp_file_directory, temp_file_name = tempfile.mkstemp(dir=os.path.dirname(zipname))
+        os.close(temp_file_directory)
+        # create a temporary copy of the archive without filename
+        with ZipFile(zipname, 'r') as zip_file_in:
+            with ZipFile(temp_file_name, 'w') as zip_file_out:
+                zip_file_out.comment = zip_file_in.comment  # preserve the comment
+                for item in zip_file_in.infolist():
+                    # keep all files that are not managed by the manager and that are not in the file bin
+                    if item.filename not in [filename, 'META-INF/manifest.xml'] + self.fileBin:
+                        zip_file_out.writestr(item, zip_file_in.read(item.filename))
+        # Replace managed file with temporary file
         os.remove(zipname)
-        os.rename(tmpname, zipname)
-
+        os.rename(temp_file_name, zipname)
         # now add filename with its new data
-        with zipfile.ZipFile(zipname, mode='a',
-                             compression=zipfile.ZIP_DEFLATED) as zf:
-            zf.writestr(filename, data)
+        with ZipFile(zipname, mode='a', compression=ZIP_DEFLATED) as zip_file:
+            zip_file.writestr('content.xml', data)
             for file in os.listdir(self.srcDir):
-                zf.write(filename=os.path.join(self.srcDir, file),
-                         arcname=os.path.join('attachments', file))
-            zf.writestr(zinfo_or_arcname='META-INF/manifest.xml',
-                        data=str(self.manifest))
-
-
-# monkey patches
-
-
-def syncOnSync(self):
-    ############################################################################
-    mapSyncer = MapSyncer(self)
-    mapSyncer.syncMaps()
-    ############################################################################
-    self.unloadCollection(self._onSync)
-
-
-sync.AnkiQt.onSync = syncOnSync
+                zip_file.write(filename=file, arcname=os.path.join('attachments', file))
+            zip_file.writestr(zinfo_or_arcname='META-INF/manifest.xml', data=str(self.manifest))
